# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Zinc Search engine API documents https://zincsearch-docs.zinc.dev",
        "title": "Zinc Search engine API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "Zinc Search",
            "url": "https://www.zincsearch.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "basePath": "/",
    "paths": {
        "/_aliases": {
            "post": {
                "security": [
                    {
                        "BasicAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Index"
                ],
                "summary": "Add or remove index alias for compatible ES",
                "operationId": "AddOrRemoveESAlias",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/meta.HTTPResponseError"
                        }
                    }
                }
            }
        },
...
        "/_index_template": {
            "get": {
                "security": [
                    {
                        "BasicAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Index"
                ],
                "summary": "List index teplates",
                "operationId": "ListTemplates",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/meta.Template"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/meta.HTTPResponseError"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BasicAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Index"
                ],
                "summary": "Create update index template",
                "operationId": "CreateTemplate",
                "parameters": [
                    {
                        "description": "Template data",
                        "name": "template",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/meta.IndexTemplate"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/meta.HTTPResponseTemplate"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/meta.HTTPResponseError"
                        }
                    }
                }
            }
        },
        "/_index_template/{name}": {
            "get": {
                "security": [
                    {
                        "BasicAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Index"
                ],
                "summary": "Get index template",
                "operationId": "GetTemplate",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Template",
                        "name": "name",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/meta.IndexTemplate"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/meta.HTTPResponseError"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BasicAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Index"
                ],
                "summary": "Create update index template",
                "operationId": "UpdateTemplate",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Template",
                        "name": "name",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Template data",
                        "name": "template",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/meta.IndexTemplate"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/meta.HTTPResponseTemplate"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/meta.HTTPResponseError"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BasicAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Index"
                ],
                "summary": "Delete template",
                "operationId": "DeleteTemplate",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Template",
                        "name": "name",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/meta.HTTPResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/meta.HTTPResponseError"
                        }
                    }
                }
            }
        },
...
        "/{target}/_alias/{target_alias}": {
            "get": {
                "security": [
                    {
                        "BasicAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Index"
                ],
                "summary": "Get index alias for compatible ES",
                "operationId": "GetESAliases",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Target Index",
                        "name": "target",
                        "in": "path"
                    },
                    {
                        "type": "string",
                        "description": "Target Alias",
                        "name": "target_alias",
                        "in": "path"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/meta.HTTPResponseError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "meta.Analyzer": {
            "type": "object",
            "properties": {
                "char_filter": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "filter": {
                    "description": "compatibility with es, alias for TokenFilter",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "lowercase": {
                    "description": "for type=pattern",
                    "type": "boolean"
                },
                "pattern": {
                    "description": "for type=pattern",
                    "type": "string"
                },
                "stopwords": {
                    "description": "for type=pattern,standard,stop",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "token_filter": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "tokenizer": {
                    "type": "string"
                },
                "type": {
                    "description": "options for compatible",
                    "type": "string"
                }
            }
        },
...
        "meta.HTTPResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
...
        "meta.HTTPResponseError": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
...
        "meta.HTTPResponseTemplate": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "template": {
                    "type": "string"
                }
            }
        },
...
        "meta.IndexAnalysis": {
            "type": "object",
            "properties": {
                "analyzer": {
                    "type": "object",
                    "additionalProperties": {
                        "$ref": "#/definitions/meta.Analyzer"
                    }
                },
                "char_filter": {
                    "type": "object",
                    "additionalProperties": true
                },
                "filter": {
                    "description": "compatibility with es, alias for TokenFilter",
                    "type": "object",
                    "additionalProperties": true
                },
                "token_filter": {
                    "type": "object",
                    "additionalProperties": true
                },
                "tokenizer": {
                    "type": "object",
                    "additionalProperties": true
                }
            }
        },
        "meta.IndexSettings": {
            "type": "object",
            "properties": {
                "analysis": {
                    "$ref": "#/definitions/meta.IndexAnalysis"
                },
                "number_of_replicas": {
                    "type": "integer"
                },
                "number_of_shards": {
                    "type": "integer"
                }
            }
        },
...
        "meta.IndexTemplate": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "index_patterns": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "priority": {
                    "description": "highest priority is chosen",
                    "type": "integer"
                },
                "template": {
                    "$ref": "#/definitions/meta.TemplateTemplate"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "meta.Mappings": {
            "type": "object",
            "properties": {
                "properties": {
                    "type": "object",
                    "additionalProperties": {
                        "$ref": "#/definitions/meta.Property"
                    }
                }
            }
        },
...
        "meta.Property": {
            "type": "object",
            "properties": {
                "aggregatable": {
                    "type": "boolean"
                },
                "analyzer": {
                    "type": "string"
                },
                "fields": {
                    "description": "Fields allow the same string value to be indexed in multiple ways for different purposes,\nsuch as one field for search and a multi-field for sorting and aggregations,\nor the same string value analyzed by different analyzers.\nIf the Fields property is defined within a sub-field, it will be ignored.\n\nCurrently, only \"text\" fields support the Fields parameter.",
                    "type": "object",
                    "additionalProperties": {
                        "$ref": "#/definitions/meta.Property"
                    }
                },
                "format": {
                    "description": "date format yyyy-MM-dd HH:mm:ss || yyyy-MM-dd || epoch_millis",
                    "type": "string"
                },
                "highlightable": {
                    "type": "boolean"
                },
                "index": {
                    "type": "boolean"
                },
                "search_analyzer": {
                    "type": "string"
                },
                "sortable": {
                    "type": "boolean"
                },
                "store": {
                    "type": "boolean"
                },
                "time_zone": {
                    "description": "date format time_zone",
                    "type": "string"
                },
                "type": {
                    "description": "text, keyword, date, numeric, boolean, geo_point",
                    "type": "string"
                }
            }
        },
...
        "meta.Template": {
            "type": "object",
            "properties": {
                "index_template": {
                    "$ref": "#/definitions/meta.IndexTemplate"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "meta.TemplateTemplate": {
            "type": "object",
            "properties": {
                "mappings": {
                    "$ref": "#/definitions/meta.Mappings"
                },
                "settings": {
                    "$ref": "#/definitions/meta.IndexSettings"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BasicAuth": {
            "type": "basic"
        }
    }
}